+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 0.0.1 - September 29th 2022
 
 - Initial release
+
+## 0.0.2 - October 12th 2022
+
+- Added a multi line plain text input widget
+- Ability to set the scrolldirection of the pageview
+- Ability to set the scrollphysics of the pages' scrollview.